+++ site/components/preset-samples/samples/mui/Sample.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { DateSelect } from "react-ymd-date-select/presets/material";
+import { DateSelect } from "react-ymd-date-select/presets/mui";
 
 function Sample() {
   const [date, setDate] = useState<string>("");
