+++ vite.config.ts
@@ -18,7 +18,7 @@ export default defineConfig({
       input: [
         path.resolve(__dirname, "index.html"),
         path.resolve(__dirname, "/samples/vanilla/index.html"),
-        path.resolve(__dirname, "/samples/material/index.html"),
+        path.resolve(__dirname, "/samples/mui/index.html"),
         path.resolve(__dirname, "/samples/chakra-ui/index.html"),
         path.resolve(
           __dirname,
